+++ lib/PuppeteerSharp/Messaging/GetLayoutMetricsResponse.cs
@@ -1,0 +1,19 @@
+﻿using Newtonsoft.Json;
+
+namespace PuppeteerSharp.Messaging
+{
+    internal class GetLayoutMetricsResponse
+    {
+        [JsonProperty("layoutViewport")]
+        public GetLayoutMetricsLayoutViewport LayoutViewport { get; set; }
+
+        internal class GetLayoutMetricsLayoutViewport
+        {
+            [JsonProperty("pageX")]
+            public decimal PageX { get; set; }
+
+            [JsonProperty("pageY")]
+            public decimal PageY { get; set; }
+        }
+    }
+}
